Centralizing the line drawing code into DrawingUtility. Fixed alignment for line numbers and normal drawing. Updated buffer position to handle top padding when calculating position.

+++ src/MfGames.GtkExt.LineTextEditor/Margins/LineNumberMarginRenderer.cs
@@ -126,12 +126,11 @@ namespace MfGames.GtkExt.LineTextEditor.Margins
 		/// <param name="lineIndex">The line index being rendered.</param>
 		/// <param name="point">The point of the specific line number.</param>
 		/// <param name="height">The height of the rendered line.</param>
-		public override void Draw(
-			IDisplayContext displayContext,
-			IRenderContext renderContext,
-			int lineIndex,
-			PointD point,
-			int height)
+		public override void Draw(IDisplayContext displayContext,
+		                          IRenderContext renderContext,
+		                          int lineIndex,
+		                          PointD point,
+		                          double height)
 		{
 			// Create a layout object if we don't have one.
 			BlockStyle style = displayContext.Theme.BlockStyles[Theme.LineNumberStyle];
